requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
openai>=1.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
libsql-experimental>=0.0.50
